@@ -1,73 +1,93 @@
 {
-  "name": "@typescord/erlpack",
-  "version": "0.3.2",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "description": "Erlpack is a fast encoder and decoder for the Erlang Term Format (version 131) for JavaScript",
-  "author": "Jason Citron",
-  "maintainers": [
-    "Typescord"
-  ],
-  "license": "MIT",
-  "engines": {
-    "node": ">=15.0.0"
-  },
-  "scripts": {
-    "install": "node-pre-gyp install --fallback-to-build",
-    "build": "node-pre-gyp install build package",
-    "build-ts": "tsc -p tsconfig.prod.json",
-    "test": "jest",
-    "lint": "eslint lib --ext ts",
-    "lint-fix": "yarn lint --fix"
-  },
-  "homepage": "https://github.com/typescord/erlpack#readme",
-  "bugs": {
-    "url": "https://github.com/typescord/erlpack/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/typescord/erlpack.git"
-  },
-  "dependencies": {
-    "@mapbox/node-pre-gyp": "^1.0.0",
-    "node-addon-api": "^3.1.0"
-  },
-  "files": [
-    "src",
-    "vendor",
-    "cpp",
-    "dist",
-    "binding.gyp"
-  ],
-  "keywords": [
-    "discord",
-    "erlpack",
-    "typescord"
-  ],
-  "devDependencies": {
-    "@types/node": "^14.14.25",
-    "@typescript-eslint/eslint-plugin": "^4.15.0",
-    "@typescript-eslint/parser": "^4.15.0",
-    "eslint": "^7.19.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "jest": "^26.6.3",
-    "prettier": "^2.2.1",
-    "ts-jest": "^26.5.1",
-    "typescript": "^4.1.4"
-  },
-  "jest": {
-    "preset": "ts-jest",
-    "testEnvironment": "node"
-  },
-  "binary": {
-    "module_name": "erlpack",
-    "module_path": "./prebuild/{node_abi}-napi-v{napi_build_version}-{platform}-{arch}-{libc}/",
-    "remote_path": "v{version}",
-    "package_name": "{module_name}-v{version}-{node_abi}-napi-v{napi_build_version}-{platform}-{arch}-{libc}.tar.gz",
-    "host": "https://github.com/typescord/erlpack/releases/download/",
-    "napi_versions": [
-      3
-    ]
-  }
+	"name": "@typescord/erlpack",
+	"version": "0.3.2",
+	"main": "dist/index.js",
+	"types": "dist/index.d.ts",
+	"description": "Erlpack is a fast encoder and decoder for the Erlang Term Format (version 131) for JavaScript",
+	"author": "Jason Citron",
+	"maintainers": [
+		"Typescord"
+	],
+	"license": "MIT",
+	"engines": {
+		"node": ">=15.0.0"
+	},
+	"scripts": {
+		"install": "node-pre-gyp install --fallback-to-build",
+		"build": "node-pre-gyp install build package",
+		"build-ts": "rimraf dist && tsc -p tsconfig.prod.json",
+		"test": "jest",
+		"lint": "eslint . --ext ts",
+		"lint-fix": "yarn lint --fix",
+		"format": "prettier . --write",
+		"mrm": "mrm --preset @typescord/mrm-preset"
+	},
+	"homepage": "https://github.com/typescord/erlpack#readme",
+	"bugs": {
+		"url": "https://github.com/typescord/erlpack/issues"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/typescord/erlpack.git"
+	},
+	"prettier": "@typescord/prettier-config",
+	"dependencies": {
+		"@mapbox/node-pre-gyp": "^1.0.0",
+		"node-addon-api": "^3.1.0",
+		"node-gyp": "^7.1.2"
+	},
+	"files": [
+		"src",
+		"vendor",
+		"cpp",
+		"dist",
+		"binding.gyp"
+	],
+	"keywords": [
+		"discord",
+		"erlpack",
+		"typescord"
+	],
+	"devDependencies": {
+		"@types/jest": "^26.0.20",
+		"@types/node": "^14.14.32",
+		"@typescord/eslint-config": "^0.1.6",
+		"@typescord/mrm-preset": "^0.1.7",
+		"@typescord/prettier-config": "^0.1.2",
+		"@typescript-eslint/eslint-plugin": "^4.16.1",
+		"@typescript-eslint/parser": "^4.16.1",
+		"eslint": "^7.21.0",
+		"eslint-config-prettier": "^8.1.0",
+		"eslint-import-resolver-node": "^0.3.4",
+		"eslint-plugin-import": "^2.22.1",
+		"eslint-plugin-prettier": "^3.3.1",
+		"eslint-plugin-simple-import-sort": "^7.0.0",
+		"eslint-plugin-sonarjs": "^0.6.0",
+		"eslint-plugin-unicorn": "^28.0.2",
+		"jest": "^26.6.3",
+		"mrm": "^2.6.0",
+		"prettier": "^2.2.1",
+		"rimraf": "^3.0.2",
+		"ts-jest": "^26.5.3",
+		"typescript": "^4.2.3"
+	},
+	"jest": {
+		"preset": "ts-jest",
+		"testEnvironment": "node"
+	},
+	"binary": {
+		"module_name": "erlpack",
+		"module_path": "./prebuild/{node_abi}-napi-v{napi_build_version}-{platform}-{arch}-{libc}/",
+		"remote_path": "v{version}",
+		"package_name": "{module_name}-v{version}-{node_abi}-napi-v{napi_build_version}-{platform}-{arch}-{libc}.tar.gz",
+		"host": "https://github.com/typescord/erlpack/releases/download/",
+		"napi_versions": [
+			3
+		]
+	},
+	"dependenciesMeta": {
+		"node-addon-api@3.1.0": {
+			"unplugged": true
+		}
+	}
 }
