@@ -1,8 +1,7 @@
 /* eslint-disable  @typescript-eslint/no-var-requires */
-import { resolve } from 'path';
-
-const { find } = require('@mapbox/node-pre-gyp');
-const { pack: _pack, unpack: _unpack } = require(find(resolve(__dirname, '..', 'package.json')));
+import { join } from 'path';
+import { find } from '@mapbox/node-pre-gyp';
+const { pack: _pack, unpack: _unpack } = require(find(join(__dirname, '../package.json')));
 
 const kPackSymbol = Symbol('erlpack.pack.custom');
 
@@ -10,14 +9,14 @@ export const packCustom = kPackSymbol;
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/explicit-module-boundary-types
 export function pack(data: any): Buffer {
-  if (data?.[kPackSymbol]) {
-    return _pack(data[kPackSymbol]());
-  }
-  return _pack(data);
+	if (data?.[kPackSymbol]) {
+		return _pack(data[kPackSymbol]());
+	}
+	return _pack(data);
 }
 export function unpack<T>(data: WithImplicitCoercion<ArrayBuffer | SharedArrayBuffer>): T {
-  if (typeof data !== 'object') {
-    throw new Error('Attempting to unpack a non-object.');
-  }
-  return _unpack(data);
+	if (typeof data !== 'object') {
+		throw new Error('Attempting to unpack a non-object.');
+	}
+	return _unpack(data);
 }
